component clean up and updated login animations
