+++ src/App.css
@@ -37,131 +37,6 @@ footer {
 article {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
-}
-
-.login-section {
-  display: flex;
-  width: 100%;
   height: 100%;
-  flex-direction: column;
-  position: relative;
-}
-
-.login-section form {
-  /* position: absolute; */
-  /* bottom: 0; */
-  display: flex;
-  height: 45%;
-  width: 100%;
-  background-color: #001524;
-  border-radius: 30px 30px 0 0;
-  flex-direction: column;
   gap: 1rem;
-  justify-content: center;
-  overflow: "auto";
-}
-
-/* .login-section form.logged-in {
-  animation-duration: 1s;
-  animation-name: slide-out;
-}
-.login-section form.logged-out {
-  animation-duration: 1s;
-  animation-name: slide-in;
-} */
-
-.login-section form button:nth-child(3) {
-  margin-top: 3rem;
-}
-
-.login-section button {
-  background-color: #15616d;
-  color: #ffecd1;
-}
-
-.logo-wrapper {
-  opacity: 0;
-  height: 55%;
-  padding: 3rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  animation-fill-mode: both;
-}
-
-.logo-wrapper.logged-in {
-  animation-name: fade-out;
-  animation-duration: 1s;
-}
-
-.logo-wrapper.logged-out {
-  animation-name: fade-in;
-  animation-duration: 3s;
-}
-
-.logo-wrapper .logo {
-  width: 50%;
-  margin-top: 6rem;
-  color: #78290f;
-}
-
-.logo-wrapper h2 {
-  color: #001524;
-  margin: 0.5rem 4rem 0.5rem 4rem;
-  text-align: center;
-  line-height: normal;
-  font-weight: 900;
-  font-size: 2em;
-}
-
-.logo-wrapper span {
-  color: #001524;
-  text-align: center;
-  margin: 0 4rem 0 4rem;
-  line-height: normal;
-}
-
-/* --- Animations --- */
-
-@keyframes slide-in {
-  from {
-    bottom: -45%;
-  }
-
-  to {
-    bottom: 0;
-  }
-}
-
-@keyframes slide-out {
-  from {
-    bottom: 0;
-  }
-
-  to {
-    bottom: -45%;
-    display: none;
-  }
-}
-
-@keyframes fade-in {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes fade-out {
-  from {
-    opacity: 1;
-  }
-
-  to {
-    opacity: 0;
-    display: none;
-  }
 }
