+++ src/components/Login/loginScreen.module.css
@@ -1,0 +1,76 @@
+.loginSection {
+  display: flex;
+  width: 100%;
+  height: 100%;
+  flex-direction: column;
+  position: relative;
+}
+
+.loginSection form {
+  display: flex;
+  height: 100vh;
+  width: 100%;
+  background-color: #001524;
+  border-radius: 30px 30px 0 0;
+  flex-direction: column;
+  gap: 1rem;
+  justify-content: flex-start;
+  position: absolute;
+  bottom: -55%;
+  left: 0;
+  right: 0;
+}
+.loginSection form button:nth-child(1) {
+  margin-top: 5rem;
+}
+
+.loginSection form button:nth-child(3) {
+  margin-top: 3rem;
+}
+
+.loginSection button {
+  background-color: #15616d;
+  color: #ffecd1;
+}
+
+.logoWrapper {
+  opacity: 0;
+  height: 55%;
+  padding: 3rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  animation-fill-mode: both;
+}
+
+.logoWrapper.logged-in {
+  animation-name: fade-out;
+  animation-duration: 1s;
+}
+
+.logoWrapper.logged-out {
+  animation-name: fade-in;
+  animation-duration: 3s;
+}
+
+.logoWrapper .logo {
+  width: 50%;
+  margin-top: 6rem;
+  color: #78290f;
+}
+
+.logoWrapper h2 {
+  color: #001524;
+  margin: 0.5rem 4rem 0.5rem 4rem;
+  text-align: center;
+  line-height: normal;
+  font-weight: 900;
+  font-size: 2em;
+}
+
+.logoWrapper span {
+  color: #001524;
+  text-align: center;
+  margin: 0 4rem 0 4rem;
+  line-height: normal;
+}
