+++ src/components/routes/HomePage/index.tsx
@@ -1,0 +1,2 @@
+import HomePage from "./HomePage"
+export default HomePage
